+++ packages/nesd/lib/ui/emulator/input/action/all_actions.dart
@@ -27,6 +27,7 @@ const emulatorActions = [
   rewindTimeline,
   decreaseVolume,
   increaseVolume,
+  screenshot,
 ];
 
 const inputActions = [