+++ packages/nesd/lib/ui/emulator/input/action_handler.g.dart
@@ -89,4 +89,4 @@ final class ActionHandlerProvider
   }
 }
 
-String _$actionHandlerHash() => r'f7872e29e8d31cdd983197c0c494b5cde2dbedd3';
+String _$actionHandlerHash() => r'ee535a6c6ce2c477effb77df19bb419bdce42fd7';