+++ packages/nesd/test/ui/emulator/input/action_handler_test.dart
@@ -7,6 +7,7 @@ import 'package:nesd/ui/emulator/input/input_action.dart';
 import 'package:nesd/ui/emulator/nes_controller.dart';
 import 'package:nesd/ui/emulator/rewind/rewind_scrub_controller.dart';
 import 'package:nesd/ui/emulator/rom_manager.dart';
+import 'package:nesd/ui/emulator/screenshot/screenshot_controller.dart';
 import 'package:nesd/ui/emulator/tools/emulator_tool.dart';
 import 'package:nesd/ui/emulator/tools/emulator_tools_controller.dart';
 import 'package:nesd/ui/emulator/tools/tool_focus_controller.dart';
@@ -29,6 +30,8 @@ class _MockToolFocusController extends Mock implements ToolFocusController {}
 
 class _MockRewindScrubController extends Mock
     implements RewindScrubController {}
+
+class _MockScreenshotController extends Mock implements ScreenshotController {}
 
 class _RecordingSink extends LogSink {
   _RecordingSink(this.records);
@@ -58,6 +61,7 @@ void main() {
       toolsController: toolsController,
       toolFocusController: _MockToolFocusController(),
       scrubController: _MockRewindScrubController(),
+      screenshotController: _MockScreenshotController(),
       actionStream: const Stream.empty(),
     );