+++ packages/nesd/test/ui/emulator/input/menu_intent_routing_test.dart
@@ -7,6 +7,7 @@ import 'package:nesd/ui/emulator/input/intents.dart';
 import 'package:nesd/ui/emulator/nes_controller.dart';
 import 'package:nesd/ui/emulator/rewind/rewind_scrub_controller.dart';
 import 'package:nesd/ui/emulator/rom_manager.dart';
+import 'package:nesd/ui/emulator/screenshot/screenshot_controller.dart';
 import 'package:nesd/ui/emulator/tools/emulator_tools_controller.dart';
 import 'package:nesd/ui/emulator/tools/tool_focus_controller.dart';
 import 'package:nesd/ui/router/router.dart';
@@ -28,6 +29,8 @@ class _MockToolFocusController extends Mock implements ToolFocusController {}
 
 class _MockRewindScrubController extends Mock
     implements RewindScrubController {}
+
+class _MockScreenshotController extends Mock implements ScreenshotController {}
 
 void main() {
   late ActionHandler handler;
@@ -56,6 +59,7 @@ void main() {
       toolsController: _MockEmulatorToolsController(),
       toolFocusController: toolFocus,
       scrubController: _MockRewindScrubController(),
+      screenshotController: _MockScreenshotController(),
       actionStream: const Stream.empty(),
     )..emulatorActive = false;