+++ CHANGELOG.md
@@ -8,6 +8,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
 ### Added
 - Color Dreams (mapper 11) support
+- Screenshot action (F12 by default) that saves the current frame 
 
 ## [0.20.0] - 2026-09-16